+++ src/Modbus/BusList.cpp
@@ -26,7 +26,39 @@
 
 using namespace Modbus;
 
+ErrorRecord::ErrorRecord() {
+    last_error_function = 0;
+    last_error_code = 0;
+    error_count = 0;
+}
+
+bool ErrorRecord::record(uint8_t func, uint8_t error) {
+    last_occurence = std::chrono::steady_clock::now();
+
+    if (last_error_function == func && error == last_error_code && error_count > 0) {
+        error_count++;
+        return false;
+    }
+    last_error_function = func;
+    last_error_code = error;
+    error_count++;
+    return true;
+}
+
+void ErrorRecord::reset() {
+    last_error_function = 0;
+    last_error_code = 0;
+    error_count = 0;
+}
+
+ErrorResponse::ErrorResponse(uint8_t address, uint8_t func)
+        : std::runtime_error(
+                  fmt::format("Error response - address {0}. response {1} ({1:02x}), function {2} ({2:02x})",
+                              address, func, func & ~BusList::MODBUS_ERROR_MASK)) {}
+
 BusList::BusList() {}
+
+int BusList::responseLength(const std::vector<uint8_t> &response) { return -1; }
 
 void BusList::parse(Parser parser) {
     auto exp_address = at(_parsed_index).buffer.address();
@@ -35,28 +67,39 @@ void BusList::parse(Parser parser) {
     uint8_t address = parser.address();
     uint8_t called = parser.func();
 
-    if (address != exp_address || called != exp_func) {
+    if (address != exp_address || (called & ~MODBUS_ERROR_MASK) != exp_func) {
         _parsed_index++;
-        throw MissingResponse(exp_address, exp_func);
+        bool new_error = _errors[exp_address].record(called, 0xff);
+
+        auto wr = WrongResponse(address, exp_address, called, exp_func);
+        if (new_error) {
+            SPDLOG_WARN(wr.what());
+        }
+        throw wr;
     }
 
-    for (auto func_record : _functions) {
-        if (func_record.func == called) {
-            func_record.action(parser);
+    if (called & MODBUS_ERROR_MASK) {
+        auto func = _functions.at(called & ~MODBUS_ERROR_MASK);
+        if (func.error_action != nullptr) {
+            func.error_action(address, called);
             _parsed_index++;
-            return;
-        }
-
-        if (func_record.error_reply == called) {
-            func_record.error_action(address, called);
+        } else {
             _parsed_index++;
-            return;
+            throw ErrorResponse(address, called);
         }
+    } else {
+        _functions.at(called).action(parser);
+        _parsed_index++;
     }
-    throw UnexpectedResponse(address, called);
+
+    // throw UnexpectedResponse(address, called);
 }
 
-void BusList::addResponse(uint8_t func, std::function<void(Parser)> action, uint8_t error_reply,
-                          std::function<void(uint8_t, uint8_t)> error_action) {
-    _functions.emplace(_functions.end(), ResponseRecord(func, action, error_reply, error_action));
+void BusList::add_response(uint8_t func, std::function<void(Parser)> action,
+                           std::function<void(uint8_t, uint8_t)> error_action) {
+    _functions.emplace(func, ResponseRecord(action, error_action));
+}
+
+void BusList::set_error_response(uint8_t func, std::function<void(uint8_t, uint8_t)> error_action) {
+    _functions.at(func).error_action = error_action;
 }